+++ src/Tokenizer/index.ts
@@ -11,9 +11,9 @@
 * file that was distributed with this source code.
 */
 
-import { EdgeError } from 'edge-error'
 import { TagStatement as BlockStatement } from '../TagStatement'
 import { MustacheStatement } from '../MustacheStatement'
+import { unclosedParen, unclosedCurlyBrace, unclosedTag } from '../Exceptions'
 
 import {
   IBlockProp,
@@ -55,7 +55,11 @@ export class Tokenizer {
   private line: number = 0
   private openedTags: IBlockNode[] = []
 
-  constructor (private template: string, private tagsDef: { [key: string]: ITagDefination }, private options: tokenizerOptions) {
+  constructor (
+    private template: string,
+    private tagsDef: { [key: string]: ITagDefination },
+    private options: tokenizerOptions,
+  ) {
   }
 
   /**
@@ -70,13 +74,12 @@ export class Tokenizer {
     }
 
     /**
-     * Process entire text, but there is an open statement, so we will
-     * process it as a raw node
+     * Done processing all the lines of the template and now we are left
+     * with a tag, which isn't properly wrapped inside `curly braces`.
+     * Maybe one or more curly braces are left opened.
      */
     if (this.blockStatement) {
-      this.consumeNode(this.getRawNode(`@${this.blockStatement.props.raw}`))
-      this.blockStatement = null
-      this.consumeNode(this.getBlankLineNode())
+      throw unclosedParen({ line: this.blockStatement.startPosition, col: 0 }, this.options.filename)
     }
 
     /**
@@ -84,24 +87,15 @@ export class Tokenizer {
      * process it as a raw node
      */
     if (this.mustacheStatement) {
-      const { raw } = this.mustacheStatement.props
-      this.mustacheStatement = null
-
-      this.consumeNode(this.getRawNode(raw))
-      this.consumeNode(this.getBlankLineNode())
+      throw unclosedCurlyBrace({ line: this.mustacheStatement.startPosition, col: 0 }, this.options.filename)
     }
 
     /**
-     * Throw exception when there are opened tags
+     * Throw exception when there are opened tags, which were never closed.
      */
     if (this.openedTags.length) {
       const openedTag = this.openedTags[this.openedTags.length - 1]
-
-      throw new EdgeError(`Unclosed tag ${openedTag.properties.name}`, 'E_UNCLOSED_TAG', {
-        line: openedTag.lineno,
-        col: 0,
-        filename: this.options.filename,
-      })
+      throw unclosedTag(openedTag.properties.name, { line: openedTag.lineno, col: 0 }, this.options.filename)
     }
   }
 
@@ -318,7 +312,10 @@ export class Tokenizer {
     const tag = this.getTag(text)
 
     /**
-     * Text is a escaped tag
+     * Text is a escaped tag, which means we need to process it as a
+     * raw node.
+     *
+     * The escaping is done to allow tags like text but aren't tags actually.
      */
     if (tag && tag.escaped) {
       this.consumeNode(this.getRawNode(text.replace(ESCAPE_REGEX, '$1')))
@@ -330,7 +327,7 @@ export class Tokenizer {
      * Text is a tag
      */
     if (tag) {
-      this.blockStatement = new BlockStatement(this.line, tag)
+      this.blockStatement = new BlockStatement(this.line, tag, this.options.filename)
       this.feedTextToBlockStatement(text.trim().replace(TRIM_TAG_REGEX, ''))
       return
     }