+++ test/statement.spec.ts
@@ -19,7 +19,7 @@ const tagDef = class If {
 
 test.group('Statement', () => {
   test('tokenize all chars of a statement', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     statement.feed('if(username)')
 
     assert.equal(statement.ended, true)
@@ -36,7 +36,7 @@ test.group('Statement', () => {
   })
 
   test('tokenize chars when it has multiple parens', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     statement.feed('if((2 + 2))')
 
     assert.equal(statement.ended, true)
@@ -52,20 +52,20 @@ test.group('Statement', () => {
   })
 
   test('throw error when trying to feed after statement has been ended', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     const fn = () => statement.feed('if((2 + 2)) then run')
-    assert.throw(fn, 'Unexpected token { then run}. Write in a new line')
+    assert.throw(fn, 'Unexpected token " then run"')
   })
 
   test('throw error when calling feed multiple times and statement is ended', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     statement.feed('if((2 + 2))')
     const fn = () => statement.feed('then run')
-    assert.throw(fn, 'Unexpected token {then run}. Write in a new line')
+    assert.throw(fn, 'Unexpected token "then run"')
   })
 
   test('parse statement into tokens when feeded in multiple lines', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     const template = dedent`if (
       2 + 2 === 4
     )`
@@ -89,7 +89,7 @@ test.group('Statement', () => {
   })
 
   test('keep ended as false when there are no parens', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     statement.feed('if')
 
     assert.equal(statement.ended, false)
@@ -105,7 +105,7 @@ test.group('Statement', () => {
   })
 
   test('keep ended as false when there are no closing parens', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     statement.feed('if(')
 
     assert.equal(statement.ended, false)
@@ -122,13 +122,13 @@ test.group('Statement', () => {
   })
 
   test('throw error when there is a closing paren without opening paren', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     const fn = () => statement.feed('if)')
-    assert.throw(fn, 'Unexpected token ). Wrap statement inside ()')
+    assert.throw(fn, 'Unexpected token ")"')
   })
 
   test('do not seek statements which are not seekable', (assert) => {
-    const statement = new TagStatement(1, Object.assign({}, tagDef, { seekable: false }))
+    const statement = new TagStatement(1, Object.assign({}, tagDef, { seekable: false }), 'welcome.edge')
     statement.feed('else')
 
     assert.equal(statement.ended, true)
@@ -144,7 +144,7 @@ test.group('Statement', () => {
   })
 
   test('trim whitespaces from statements which are not seekable', (assert) => {
-    const statement = new TagStatement(1, Object.assign({}, tagDef, { seekable: false }))
+    const statement = new TagStatement(1, Object.assign({}, tagDef, { seekable: false }), 'welcome.edge')
     statement.feed('  else  ')
 
     assert.equal(statement.ended, true)
@@ -160,7 +160,7 @@ test.group('Statement', () => {
   })
 
   test('record whitespaces for multi line statements', (assert) => {
-    const statement = new TagStatement(1, tagDef)
+    const statement = new TagStatement(1, tagDef, 'welcome.edge')
     const template = dedent`if(
       users.find((user) => {
         return user.username === 'virk'
@@ -184,7 +184,7 @@ test.group('Statement', () => {
   })
 
   test('set selfclosed to true for when bang is detected', (assert) => {
-    const statement = new TagStatement(1, Object.assign({}, tagDef, { selfclosed: true }))
+    const statement = new TagStatement(1, Object.assign({}, tagDef, { selfclosed: true }), 'welcome.edge')
     statement.feed('!each(user in users)')
 
     assert.equal(statement.ended, true)