+++ test/tokenizer-mustache.spec.ts
@@ -10,6 +10,7 @@
 import * as test from 'japa'
 import * as dedent from 'dedent'
 import { Tokenizer } from '../src/Tokenizer'
+import { NodeType } from '../src/Contracts'
 
 const tagsDef = {
   if: class If {
@@ -40,12 +41,12 @@ test.group('Tokenizer Mustache', () => {
     assert.isNull(tokenizer['mustacheStatement'])
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'Hello ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -54,7 +55,7 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 1,
       },
     ])
@@ -70,12 +71,12 @@ test.group('Tokenizer Mustache', () => {
     assert.isNull(tokenizer['mustacheStatement'])
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'Hello ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -84,12 +85,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: '!',
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 1,
       },
     ])
@@ -109,12 +110,12 @@ test.group('Tokenizer Mustache', () => {
     assert.isNull(tokenizer['mustacheStatement'])
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'List of users are ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -123,12 +124,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 5,
         value: '.',
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 5,
       },
     ])
@@ -148,12 +149,12 @@ test.group('Tokenizer Mustache', () => {
     assert.isNull(tokenizer['mustacheStatement'])
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'List of users are ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 's__mustache',
@@ -162,12 +163,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 5,
         value: '.',
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 5,
       },
     ])
@@ -184,12 +185,12 @@ test.group('Tokenizer Mustache', () => {
 
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'Hello ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -198,12 +199,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ', your age is ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -212,7 +213,7 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 1,
       },
     ])
@@ -236,12 +237,12 @@ test.group('Tokenizer Mustache', () => {
 
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'Hello ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -250,12 +251,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ', your friends are ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -264,45 +265,37 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 5,
         value: '!',
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 5,
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 6,
         value: 'Bye',
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 6,
       },
     ])
   })
 
-  test('convert incomplete mustache statements to raw string', (assert) => {
+  test('raise error if incomplete mustache statements', (assert) => {
+    assert.plan(2)
     const template = 'Hello {{ username'
 
     const tokenizer = new Tokenizer(template, tagsDef, { filename: 'foo.edge' })
-    tokenizer.parse()
-
-    assert.isNull(tokenizer['blockStatement'])
-    assert.isNull(tokenizer['mustacheStatement'])
-    assert.deepEqual(tokenizer.tokens, [
-      {
-        type: 'raw',
-        lineno: 1,
-        value: 'Hello {{ username',
-      },
-      {
-        type: 'newline',
-        lineno: 1,
-      },
-    ])
+    try {
+      tokenizer.parse()
+    } catch ({ message, line }) {
+      assert.equal(message, 'Missing token "}"')
+      assert.equal(line, 1)
+    }
   })
 
   test('parse 3 mustache statements in a single line', (assert) => {
@@ -316,7 +309,7 @@ test.group('Tokenizer Mustache', () => {
 
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -325,12 +318,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ', ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -339,12 +332,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ' and ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -353,7 +346,7 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 1,
       },
     ])
@@ -370,7 +363,7 @@ test.group('Tokenizer Mustache', () => {
 
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -379,12 +372,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ', ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'e__mustache',
@@ -393,7 +386,7 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 1,
       },
     ])
@@ -412,7 +405,7 @@ test.group('Tokenizer Mustache', () => {
 
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -421,12 +414,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ', ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'e__mustache',
@@ -435,7 +428,7 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 3,
       },
     ])
@@ -452,12 +445,12 @@ test.group('Tokenizer Mustache', () => {
 
     assert.deepEqual(tokenizer.tokens, [
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: 'Hello ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'e__mustache',
@@ -466,12 +459,12 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'raw',
+        type: NodeType.RAW,
         lineno: 1,
         value: ', your age is ',
       },
       {
-        type: 'mustache',
+        type: NodeType.MUSTACHE,
         lineno: 1,
         properties: {
           name: 'mustache',
@@ -480,31 +473,23 @@ test.group('Tokenizer Mustache', () => {
         },
       },
       {
-        type: 'newline',
+        type: NodeType.NEWLINE,
         lineno: 1,
       },
     ])
   })
 
-  test('use raw text when mustache is not closed properly', (assert) => {
+  test('raise error if mustache is not properly closed', (assert) => {
+    assert.plan(2)
+
     const template = dedent`Hello {{ username }.`
-
     const tokenizer = new Tokenizer(template, tagsDef, { filename: 'foo.edge' })
-    tokenizer.parse()
 
-    assert.isNull(tokenizer['blockStatement'])
-    assert.isNull(tokenizer['mustacheStatement'])
-
-    assert.deepEqual(tokenizer.tokens, [
-      {
-        type: 'raw',
-        lineno: 1,
-        value: 'Hello {{ username }.',
-      },
-      {
-        type: 'newline',
-        lineno: 1,
-      },
-    ])
+    try {
+      tokenizer.parse()
+    } catch ({ message, line }) {
+      assert.equal(message, 'Missing token "}"')
+      assert.equal(line, 1)
+    }
   })
 })